ui: Merge timeline::room_ext module into timeline::traits
