--- crates/matrix-sdk-ui/src/timeline/room_ext.rs
@@ -1,35 +1,0 @@
-// Copyright 2023 The Matrix.org Foundation C.I.C.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use async_trait::async_trait;
-use matrix_sdk::room;
-
-use crate::timeline::Timeline;
-
-#[async_trait]
-pub trait RoomExt {
-    /// Get a [`Timeline`] for this room.
-    ///
-    /// This offers a higher-level API than event handlers, in treating things
-    /// like edits and reactions as updates of existing items rather than new
-    /// independent events.
-    async fn timeline(&self) -> Timeline;
-}
-
-#[async_trait]
-impl RoomExt for room::Common {
-    async fn timeline(&self) -> Timeline {
-        Timeline::builder(self).track_read_marker_and_receipts().build().await
-    }
-}
